+++ .eslintrc.js
@@ -61,5 +61,6 @@ module.exports = {
     '@sentry-internal/sdk/no-async-await': 'off',
     '@sentry-internal/sdk/no-optional-chaining': 'off',
     '@sentry-internal/sdk/no-nullish-coalescing': 'off',
+    '@sentry-internal/sdk/no-unsupported-es6-methods': 'off',
   },
 };